crud javítások

+++ logicals/torol.php
@@ -1,8 +1,7 @@
 <?php
-$dbh = new PDO('mysql:host=localhost;dbname=gyakorlat7','root','');
-
-$stmt = $dbh->prepare("DELETE FROM np WHERE id = ?");
-$stmt->execute([$_GET['id']]);
-
-header("Location: index.php?tablazat");
+if (isset($_GET['id'])) {
+    $stmt = $dbh->prepare("DELETE FROM np WHERE id = ?");
+    $stmt->execute([$_GET['id']]);
+}
+header("Location: ./?oldal=tablazat");
 exit;
